Backend/requirements.txt
fastapi==0.110.1
fastapi-socketio==0.0.10
uvicorn==0.25.0
openapi-schema-pydantic==1.2.4
pydantic==2.7.0
pydantic_core==2.18.1
langchain==0.1.16
langchain-community==0.0.32
langchain-core==0.1.42
langchain-openai==0.0.2.post1
langchain-text-splitters==0.0.1
python-dotenv==1.0.1
